+++ packages/cli/CHANGELOG.md
@@ -29,8 +29,12 @@
   - Root-level tasks will now create a root `moon.yml`, instead of warning.
 - Added `unixShell` and `windowsShell` task options, so that the underlying shell can be configured
   per task.
+- Added `bun.inferTasksFromScripts` setting to `.moon/toolchain.yml`, for compatibility with
+  Node.js.
 - Added environment variable support to `fileGroups`.
 - Added a `@envs(group)` token function for referencing environment variables.
+- Deprecated the `moon node run-script` command. Run the task through a package manager instead,
+  like `npm run` or `yarn run`.
 - Updated tasks with glob-like arguments to automatically enabled the `shell` option, so that glob
   expansion works correctly.
 - Implemented a new console layer for writing to stdout/stderr.